refactor: Move getmembership specifications to Utils
